ref: 코드리뷰 후 코드 보완
