+++ src/main/kotlin/calculator/CalculatorApplication.kt
@@ -1,0 +1,6 @@
+package calculator
+
+fun main() {
+    val input = readLine()
+    println(StringAddCalculator().add(input))
+}
